Revert "Windows Opengl context"
