+++ .vscode/settings.json
@@ -2,9 +2,6 @@
     "files.associations": {
         "camera": "c",
         "graphicscontext": "c",
-        "viewport": "c",
-        "ma-window.h": "c",
-        "wgl.h": "c",
-        "stdlib.h": "c"
+        "viewport": "c"
     }
 }
