+++ sandbox/CMakeLists.txt
@@ -1,9 +1,3 @@
 add_executable(sandbox src/main.c)
 
-find_package(OpenGL REQUIRED)
-
-target_link_libraries(sandbox PRIVATE ma-window ${OPENGL_gl_LIBRARY})
-
-if(WIN32)
-    target_compile_definitions(sandbox PRIVATE WIN32)
-endif()
+target_link_libraries(sandbox PRIVATE ma-window GL)
